Writing a Smart Contract in Ethereum
6.2 Demonstrating the infrastructures and the Solidity code
The infrastructures of components 1a and 1b have to be set up by the students as a group before the assessment (component 5) starts. The infrastructure of component 3 has to be set up by the student individually before the assessment. The Solidity code of component 2 has to be on the infrastructure of component 1b before the assessment starts.